Fix some errors in xkb
The generated code did not compile. Example build errors are:

error[E0308]: mismatched types
    --> x11rb-protocol/src/protocol/xkb.rs:6635:66
     |
6635 |         let details_bytes = self.details.serialize(affect_which, self.clear, self.select_all);
     |                                                                  ^^^^^^^^^^ expected `u16`, found struct `EventType`

error[E0308]: mismatched types
    --> x11rb-protocol/src/protocol/xkb.rs:6635:78
     |
6635 |         let details_bytes = self.details.serialize(affect_which, self.clear, self.select_all);
     |                                                                              ^^^^^^^^^^^^^^^ expected `u16`, found struct `EventType`

This commit introduces mostly useless conversions to work around this problem.

+++ x11rb-protocol/src/protocol/render.rs
@@ -1814,7 +1814,7 @@ impl CreatePictureAux {
     #[allow(dead_code)]
     fn serialize(&self, value_mask: u32) -> Vec<u8> {
         let mut result = Vec::new();
-        self.serialize_into(&mut result, value_mask);
+        self.serialize_into(&mut result, u32::from(value_mask));
         result
     }
     fn serialize_into(&self, bytes: &mut Vec<u8>, value_mask: u32) {
@@ -2032,7 +2032,7 @@ impl<'input> CreatePictureRequest<'input> {
             value_mask_bytes[3],
         ];
         let length_so_far = length_so_far + request0.len();
-        let value_list_bytes = self.value_list.serialize(value_mask);
+        let value_list_bytes = self.value_list.serialize(u32::from(value_mask));
         let length_so_far = length_so_far + value_list_bytes.len();
         let padding0 = &[0; 3][..(4 - (length_so_far % 4)) % 4];
         let length_so_far = length_so_far + padding0.len();
@@ -2220,7 +2220,7 @@ impl ChangePictureAux {
     #[allow(dead_code)]
     fn serialize(&self, value_mask: u32) -> Vec<u8> {
         let mut result = Vec::new();
-        self.serialize_into(&mut result, value_mask);
+        self.serialize_into(&mut result, u32::from(value_mask));
         result
     }
     fn serialize_into(&self, bytes: &mut Vec<u8>, value_mask: u32) {
@@ -2426,7 +2426,7 @@ impl<'input> ChangePictureRequest<'input> {
             value_mask_bytes[3],
         ];
         let length_so_far = length_so_far + request0.len();
-        let value_list_bytes = self.value_list.serialize(value_mask);
+        let value_list_bytes = self.value_list.serialize(u32::from(value_mask));
         let length_so_far = length_so_far + value_list_bytes.len();
         let padding0 = &[0; 3][..(4 - (length_so_far % 4)) % 4];
         let length_so_far = length_so_far + padding0.len();
